New class attestations added for logic and added missing test attestations file
