+++ test/fixtures/profiles/attestation/files/no-status.yaml
@@ -1,0 +1,4 @@
+tmp-3.0.3:
+  justification: Unassailable thinking
+  evidence_url: Dummy url
+  expiration_date: 2050-06-01
